+++ jetstream/analysis.py
@@ -19,34 +19,10 @@ from mozanalysis.utils import add_days
 from . import AnalysisPeriod
 from jetstream.config import AnalysisConfiguration
 from jetstream.dryrun import dry_run_query
+import jetstream.errors as errors
 from jetstream.statistics import Count, StatisticResult, StatisticResultCollection
 
 logger = logging.getLogger(__name__)
-
-
-class NoSlugException(Exception):
-    def __init__(self, message="Experiment has no slug"):
-        super().__init__(message)
-
-
-class NoEnrollmentPeriodException(Exception):
-    def __init__(self, normandy_slug, message="Experiment has no enrollment period"):
-        super().__init__(f"{normandy_slug} -> {message}")
-
-
-class NoStartDateException(Exception):
-    def __init__(self, normandy_slug, message="Experiment has no start date."):
-        super().__init__(f"{normandy_slug} -> {message}")
-
-
-class EndedException(Exception):
-    def __init__(self, normandy_slug, message="Experiment has already ended."):
-        super().__init__(f"{normandy_slug} -> {message}")
-
-
-class EnrollmentLongerThanAnalysisException(Exception):
-    def __init__(self, normandy_slug, message="Enrollment period is longer than analysis dates."):
-        super().__init__(f"{normandy_slug} -> {message}")
 
 
 @attr.s(auto_attribs=True)
@@ -127,7 +103,7 @@ class Analysis:
         )
 
         if analysis_length_dates < 0:
-            raise EnrollmentLongerThanAnalysisException(self.config.experiment.normandy_slug)
+            raise errors.EnrollmentLongerThanAnalysisException(self.config.experiment.normandy_slug)
 
         return TimeLimits.for_single_analysis_window(
             last_date_full_data=prior_date_str,
@@ -287,20 +263,20 @@ class Analysis:
     def check_runnable(self, current_date: Optional[datetime] = None) -> bool:
         if self.config.experiment.normandy_slug is None:
             # some experiments do not have a normandy slug
-            raise NoSlugException()
+            raise errors.NoSlugException()
 
         if not self.config.experiment.proposed_enrollment:
-            raise NoEnrollmentPeriodException(self.config.experiment.normandy_slug)
+            raise errors.NoEnrollmentPeriodException(self.config.experiment.normandy_slug)
 
         if self.config.experiment.start_date is None:
-            raise NoStartDateException(self.config.experiment.normandy_slug)
+            raise errors.NoStartDateException(self.config.experiment.normandy_slug)
 
         if (
             current_date
             and self.config.experiment.end_date
             and self.config.experiment.end_date < current_date
         ):
-            raise EndedException(self.config.experiment.normandy_slug)
+            raise errors.EndedException(self.config.experiment.normandy_slug)
 
         return True
